+++ backend/src/game/systems/area_controller.rs
@@ -1,5 +1,5 @@
 use shared::{
-    constants::MAX_AREA_LEVEL,
+    constants::{MAX_AREA_LEVEL, MAX_POWER_SHARD_LEVEL_BASE},
     data::{
         area::{AreaLevel, AreaSpecs, AreaState},
         item::ItemSpecs,
@@ -72,6 +72,12 @@ pub fn init_area_specs(
     area_specs.effects = EffectsMap::combine_all(
         std::iter::once(map_effects).chain(std::iter::once(area_specs.effects)),
     );
+
+    area_specs.max_power_shards_level = map_item
+        .as_ref()
+        .and_then(|map_item| map_item.map_specs.as_ref())
+        .and_then(|map_specs| map_specs.max_power_shards_level)
+        .unwrap_or(MAX_POWER_SHARD_LEVEL_BASE);
 
     compute_area_specs(&mut area_specs);