+++ backend/src/game/systems/events_resolver.rs
@@ -2,7 +2,9 @@ use std::iter;
 
 use itertools::all;
 use shared::{
-    constants::{ITEM_REWARDS_BOSS_FACTOR, MAX_AREA_LEVEL, WAVES_PER_AREA_LEVEL},
+    constants::{
+        ITEM_REWARDS_BOSS_FACTOR, MAX_AREA_LEVEL, POWER_SHARD_LEVELS_NEEDED, WAVES_PER_AREA_LEVEL,
+    },
     data::{
         area::{AreaLevel, ThreatLevel},
         character::CharacterId,
@@ -462,9 +464,10 @@ fn handle_area_completed_event(
     let area_state = game_data.area_state.mutate();
 
     if game_data.area_specs.can_reward_shards()
-        && (area_state.area_level > area_state.max_area_level_ever)
-        && (area_state.area_level).is_multiple_of(10)
+        && (area_state.area_level > area_state.max_power_shard_level_ever)
+        && (area_state.area_level).is_multiple_of(POWER_SHARD_LEVELS_NEEDED)
     {
+        area_state.max_power_shard_level_ever = area_level;
         game_data.player_resources.mutate().shards += 1.0;
     }