+++ shared/src/data/user.rs
@@ -57,4 +57,5 @@ pub struct UserCharacter {
 pub struct UserGrindArea {
     pub area_id: String,
     pub max_level_reached: AreaLevel,
+    pub max_power_shard_level: AreaLevel,
 }